+++ src/views/Sales/components/SaleFinalPrice/index.spec.tsx
@@ -8,7 +8,7 @@ import '@testing-library/jest-dom/extend-expect'
 import { SaleFinalPrice } from 'src/views/Sales/components/SaleFinalPrice/index'
 
 // utils
-import { getSaleDefault, addHours, dateUTC } from 'src/utils/helpers'
+import { getSaleDefault, addHours, dateUTC } from 'src/utils'
 
 // theme
